Move Tag enum to TiffImage.hpp (Fixes #9)
- Move Tag enum to TiffImage.hpp
- Move `toString` functions to TiffStdOut.hpp
- Delete TiffTags.hpp
